requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
PyPDF2
python-dotenv
email-validator
dateparser
google-generativeai